Remove MainOpt constructor and add init
The presence of the constructor causes the file writer not to write data
to file in the integration test (the command is received, as the NeXus
file is created with the correct structure). Just commenting out the
contents of the constructor makes no difference. As a temporary
solution, an init function was added to initialise the variable that was
being initialised in the constructor. If the cause of the problem is
discovered and fixed, the constructor can be reintroduced.