Ignore broker addresses sent in JobStart message
We want to decouple the Kafka addresses seen by Nicos and the Filewriter.
In this commit we stop honouring the address sent by Nicos and use one
of the addresses already known by the Filewriter.

In the future we want to refactor the Kafka configuration in the filewriter,
possibly resulting in a single "kafka-config" that the code in this commit
will use to get the broker address from, instead of using JobPoolURI.
See ECDC-3118.