Fix: Do not require service_id on run_stop messages
The field is optional in the 6s4t schema.

In the context of this PR, it is necessary to fix this issue because when
stopping a job immediately after start, whoever started the job will
probably not have information about which filewriter has picked up the
job, and may therefore send the 6s4t message with an empty service_id.