Feat(metadata): Write units attribute in derived datasets (metadata)
We modify Value class to also receive a function to write attributes
and assume that Value represents a dataset to be written with or without
attributes.
Previously, Value could be called separately with a group/dataset and
a function to respectively create a dataset in the group / attribute
in the dataset.
I do not see the usefulness of such generalisation, Value class is used
exclusively to create derived datasets, so it seems simpler to model it
as a dataset and nothing else.