Fix: Command with invalid IDs are silently skipped instead of producing a failed command response
Ever since we stopped skipping commands received without an explicit
kafka-to-nexus `service_id`, filewriters listening in the same topic
as the intended recipient of a Command message will issue error
command responses to the command.
Our clients do not handle this error messages gracefully, this commit
attempts to fix the issue on the filewriter side by preventing
some of those error responses.