Fix: Stop commands should return errors only if service_id matches
ServiceID is optional in stop commands. PR #676 leveraged this
to start accepting stop commands with empty ServiceID
(see b33cbafbc54a4b860f4ed37b3c879a4bc5081e25).

This in turn caused sibling filewriters to respond to those stop
commands, as it was no longer clear if the message was addressed
to them (see logic in Handler.cpp).
Furthermore, the response would typically be an error response,
as those filewriters would have no knowledge about the job that
is being stopped.

Those error responses may mislead the entity sending the command,
so this commit changes the response behaviour to only send error
responses if ServiceID was not empty and matches with the
ServiceID of the receiving filewriter.