Refa: Actually implement Command::Handler constructor with depend. injection
The constructor that uses dep. injection did not actually have an implementation.
Moreover, the signature had to be changed, as class sometimes replaces
the injected objects by new ones, so we need to know the Kafka URIs and
Settings to recreate new objects.

(this shows that the class needs a bigger refactor, but we will try to
make some progress for now to at least allow unit testing)