[Refactor] timestamp NeXusDataset classes
Remove redundancy added by introducing Signed* timestamp classes *and*
present previously between Time and CueTimestampZero.

Time and CueTimestampZero are now each specializations of an
intermediate templated class that inserts the 'correct' dataset name for
a Timestamp class that always has start=1970-01-01T00:00:00Z and
units=ns attributes.
The specializations could be removed if every use of Time and
CueTimestampZero throughout kafka-to-nexus' modules was replaced by
TypedTime<> and TypedCueTimestampZero<>, respectively; both of which
default to the now-disfavoured std::uint64_t datatype.
Following that, the intermediate templated classes could be renamed,
dropping the 'Typed' prefix.

This change makes the timesamp constructor a templated constructor, so
adds (slightly?) to compile-time for every compilation unit that needs
either Time or CueTimestampZero.